requirements.txt
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
# IARA Ecosystem — Dependências Python
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

# ── FastAPI + Servidor ──────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# ── Telegram Bot ────────────────────────────────────────
python-telegram-bot>=21.0

# ── HTTP Client (Assíncrono) ────────────────────────────
aiohttp>=3.10.0
httpx>=0.27.0

# ── Banco de Dados ──────────────────────────────────────
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.35

# ── Redis ───────────────────────────────────────────────
redis[hiredis]>=5.2.0

# ── Qdrant ──────────────────────────────────────────────
qdrant-client>=1.12.0

# ── LLM Router ──────────────────────────────────────────
litellm>=1.52.0

# ── Memória Inteligente ─────────────────────────────────
mem0ai>=0.1.30

# ── LightRAG (Graph-augmented RAG) ──────────────────────
lightrag-hku>=1.0.0

# ── LangGraph (Orquestração Macro) ──────────────────────
langgraph>=0.2.50

# ── MCP SDK (Model Context Protocol) ───────────────────
mcp>=1.2.0

# ── Docker SDK (Sandbox Arena) ──────────────────────────
docker>=7.1.0

# ── Observabilidade ─────────────────────────────────────
sentry-sdk[fastapi]>=2.18.0

# ── Utilidades ──────────────────────────────────────────
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.6.0
